Example test to show issue with rasterio reading metadata tags
This commit is not meant to be merged!
Provides an example unit test to show an issue with reading tags from
a GeoTIFF with malformated characters. Requires installing GDAL python
bindings to run. The test shows that GDAL python binding handles
reading the metadata from the GeoTiff without error, while rasterio
throws a UnicodeDecodeError.